mongoose: add native and message based WebSocket ping options
A WebSocket connection may get disconnected at any time, without the
client getting notified. Using native WebSocket Ping frames is common
method keeping a connection open.
Furthermore, when using SWUpdate behind a reverse proxy, connections
are usually dropped after an idle timeout. Mostly within one minute!

The SWUpdate wep-app automatically reloads, if the WebSocket connection
is closed. This leads to a confusing message to the user, that the
device is being restarted.

As web browsers don't support sending Ping frames, it's advisable to
send them from the server to keep the connection open.

By default, this feature is disabled and needs to be enabled with the
new mongoose setting `ws_ping_interval`. The value corresponds to the
ping interval in seconds.